Added an upgrade to convert current metadata to the new system
Changed item schema so that it's columns match the metadata columns.

Still haven't hooked up this stuff yet, but I think we're finally ready.

Made hopefully the final version of the sqlite rename/delete column trick.
This one does it all.

Fixed upgrading databases in the unittests.  Before if we would fail then we
would try to throw up a modal dialog which is not so hot when trying to run
automated unittests.