Performance improvements for metadata system
Avoid trying to write metadata while we're initially adding the files.  We
want to write things out to sqlite as fast as possible and reading things from
mutagen hurts that.

Don't create an idle for each TaskResult.  When we're running mutagen we
create dozens of results per second.  If we create idles for each one of them,
we can clog up the idle loop.